+++ CHANGELOG.md
@@ -1,3 +1,27 @@
+# version 11.02
+
+Added the option to set the image and colour specific to the type of creature.
+
+Added blob, scorch marks, corpses, and bones to the list of images that could be used.
+
+Also added none in case you just want the token to completely disappear.
+
+And unconscious in case you want the token to remain with no images.
+
+Added a setting to remove any animation effects when a token is defeated.
+
+Added the option to disable any tokens that have been defeated in combat so they don't accidentally get in the way of targeting.
+
+Fixed issue where you'd have to click twice to get the image to restore after healing an NPC.
+
+Fixed issue with the fade animation when restoring an NPC.
+
+Fixed issue in PF2E when a combat hasn't started yet.
+
+Fixed issue when loading Foundry or loading a scene where the token wasn't refreshing to the correct image.
+
+Added updated chinese translations
+
 # version 11.01
 
 Adding v11 support